confirm_cleaning_history

+++ assets/javascript/index.js
@@ -144,10 +144,15 @@ class ViewLayer {
 
   clearHistory() {
     this.elements.clearButton.addEventListener("click",  (e) => {
-      localStorage.setItem("reposlist", JSON.stringify([]));
-      instance_of_view.elements["repo_list"].innerHTML = ""; })
-  }
+    let clearCashe = confirm("are you sure want to clear the cache");
+    if ( clearCashe === true) {
+     localStorage.setItem('reposlist', JSON.stringify([]));
+      instance_of_view.elements["repo_list"].innerHTML = "";
+       }
+    });  
+ }
 
+ 
   render(data){
     if(debug){console.log("userprofile got updated")}
     instance_of_view.elements.image.src = data.avatar_url
